Authentication using JWT

+++ controllers/api/v1/post_api.js
@@ -21,13 +21,21 @@ module.exports.index=async function(req,res){
 module.exports.destroy=async function(req,res){
 
     try{
+        console.log(req.params.id);
         let post=await Post.findById(req.params.id);
+        if(post.user==req.user.id){
+            post.remove();
+            await Comment.deleteMany({post:req.params.id});
+                
+            return res.json(200,{
+                message:'Post and associated comments deleted'
+            });
+        }else{
+            return res.json(401,{
+                message:'You cannot delete this post'
+            })
+        }
         
-        post.remove();
-        await Comment.deleteMany({post:req.params.id});
-        return res.json(200,{
-            message:'Post and associated comments deleted'
-        })
     }catch(err){
         return res.json(500,{
             message:'Internal Server Error'
